Refactor file locations for cgo build
