+++ .gitignore
@@ -1,6 +1,6 @@
 client/dist
 client/node_modules
 
-rpicam/_build
+rpicam/native/_build
 
 baby-monitor
